requirements.txt
asgiref==3.5.2
autopep8==1.5.7
blis==0.7.9
catalogue==2.0.8
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
confection==0.0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
cymem==2.0.7
defusedxml==0.7.1
distlib==0.3.6
Django==4.1.1
django-cors-headers==3.13.0
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.1/en_core_web_sm-3.4.1-py3-none-any.whl
filelock==3.8.0
flake8==3.9.2
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
langcodes==3.3.0
MarkupSafe==2.1.1
mccabe==0.6.1
murmurhash==1.0.9
numpy==1.23.4
oauthlib==3.2.1
packaging==21.3
pathy==0.6.2
platformdirs==2.5.2
preshed==3.0.8
pycodestyle==2.7.0
pycparser==2.21
pydantic==1.10.2
pyflakes==2.3.1
PyJWT==2.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
smart-open==5.2.1
social-auth-app-django==4.0.0
social-auth-core==4.3.0
spacy==3.4.2
spacy-legacy==3.0.10
spacy-loggers==1.0.3
sqlparse==0.4.3
srsly==2.4.5
thinc==8.1.5
toml==0.10.2
tqdm==4.64.1
typer==0.4.2
typing-extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
virtualenv==20.17.1
wasabi==0.10.1
